% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{attach_draw}
\alias{attach_draw}
\alias{attach_sample}
\title{Attach a draw or sample to an environment}
\usage{
attach_draw(fit, index, chain, iter, envir = parent.frame())

attach_sample(fit, index, chain, iter, envir = parent.frame())
}
\arguments{
\item{fit}{A Bayesian Fit or a Draw Data Frame as generated by \code{\link[=get_draws]{get_draws()}}.}

\item{index}{Integer indicating the row in \code{draws} to attach. Alternatively, you can
specify \code{chain} and \code{iter}.}

\item{chain}{Integer indicating the chain.}

\item{iter}{Integer indicating the within-chain iteration.}

\item{envir}{Environment, where to attach the draw. See \code{\link[base:list2env]{base::list2env()}}.}
}
\value{
The draw as a list, invisibly.
}
\description{
Attach a draw or sample to an environment
}
\examples{
for (i in 1:5) {
    attach_draw(jags_fit, index = i)
    print(beta)
}
}
