% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{cmdstan_model}
\alias{cmdstan_model}
\title{Compile a Model with CmdStan}
\usage{
cmdstan_model(file = NULL, quiet = FALSE, ...)
}
\arguments{
\item{file}{String indicating the file containing the model. Can either be
a .exe file with the compiled model or a .stan file containing the model code.}

\item{quiet}{Logical, wether to suppress messages.}

\item{...}{Other arguments passed on to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A CmdStanModel object.
}
\description{
Compile a Model with CmdStan
}
