% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_lp}
\alias{get_lp}
\title{Get the Log-Probability of an optimized fit.}
\usage{
get_lp(fit)
}
\arguments{
\item{fit}{Fit optimized with \code{\link[=run_cmdstan_optimizer]{run_cmdstan_optimizer()}}.}
}
\description{
Get the Log-Probability of an optimized fit.
}
