% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_means}
\alias{get_means}
\title{Get posterior means.}
\usage{
get_means(fit)
}
\arguments{
\item{fit}{STAN or JAGS fit.}
}
\value{
Named list, where the names are the parameters.
}
\description{
Get the posterior means from a bayesian fit.
}
