% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_means}
\alias{get_means}
\alias{get_sd}
\title{Get posterior means and standard deviations.}
\usage{
get_means(fit)

get_sd(fit)
}
\arguments{
\item{fit}{Bayesian Fit.}
}
\value{
Named list, where the names are the parameters.
}
\description{
Get the posterior means from a bayesian fit.
}
\seealso{
[get_statistic()]
}
