% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_mle}
\alias{get_mle}
\title{Get the Maximum Likelihood Estimation of an optimized fit.}
\usage{
get_mle(fit)
}
\arguments{
\item{fit}{Fit optimized with \code{\link[=run_cmdstan_optimizer]{run_cmdstan_optimizer()}}.}
}
\description{
Get the Maximum Likelihood Estimation of an optimized fit.
}
