% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_parameter_dim}
\alias{get_parameter_dim}
\title{Get dimensions of parameters.}
\usage{
get_parameter_dim(fit)
}
\arguments{
\item{x}{A Bayesian Fit or a Draw Data Frame as generated from \code{\link[=get_draws]{get_draws()}}.}
}
\description{
Get dimensions of parameters.
}
