% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_parameters}
\alias{get_parameters}
\title{Get the names of parameters as a character vector.}
\usage{
get_parameters(fit)
}
\arguments{
\item{fit}{A Bayesian Fit or a Draw Data Frame as generated from \code{\link[=get_draws]{get_draws()}}.}
}
\description{
Get the names of parameters as a character vector.
}
