% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_draws}
\alias{get_draws}
\alias{get_samples}
\title{Get posterior samples}
\usage{
get_draws(fit, as = c("df", "list", "array"))

get_samples(fit, as = c("df", "list", "array"))
}
\arguments{
\item{fit}{STAN or JAGS fit.}

\item{as}{String, "df" to return a tibble with iteration and chain,
"list" to return a list of parameters, and
"array" to return a 3-D array (iter, chain, parameter).}
}
\description{
Get the posterior samples from a bayesian fit.
}
