% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{get_samples}
\alias{get_samples}
\alias{get_posterior}
\title{Get posterior samples}
\usage{
get_samples(x, n = NULL, ...)

get_posterior(x, n = NULL, ...)
}
\arguments{
\item{x}{STAN or JAGS fit.}

\item{n}{Number of samples to extract.}
}
\description{
Get the posterior samples from a bayesian fit.
}
