% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_statistic.default}
\alias{get_statistic.default}
\title{Get an arbitrary statistic from each parameter of a fit.}
\usage{
\method{get_statistic}{default}(fit, fun = mean)
}
\arguments{
\item{fit}{Bayesian fit.}

\item{fun}{Function to apply to each parameter.}
}
\description{
Get an arbitrary statistic from each parameter of a fit.
}
