% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jags_fit}
\alias{jags_fit}
\title{Example JAGS fit}
\format{
Fit with 3 parameters:
\itemize{
\item alpha (scalar)
\item beta (vector of length 3)
\item gamma (matrix of size 3x2)
}
}
\usage{
jags_fit
}
\description{
Example JAGS fit
}
\keyword{datasets}
