% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{jags_model}
\alias{jags_model}
\title{Write a JAGS model with R's syntax highlighting.}
\usage{
jags_model(model, data)
}
\description{
Instead of writing the model on a string, write it inside
this function and get syntax highlighting and warnings.
}
\examples{
my_model <- jags_model({
    a ~ dexp(0.01)
    b ~ dexp(0.01)

    for (i in 1:n) {
        lambda[i] ~ dgamma(a, b)
        y[i] ~ dpois(lambda[i])
    }
})
}
