% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{map_draws}
\alias{map_draws}
\title{Apply a function to each draw.}
\usage{
map_draws(fit, fun)
}
\arguments{
\item{fit}{A Bayesian Fit or a Draw Data Frame as generated by \code{\link[=get_draws]{get_draws()}}.}

\item{fun}{Function that takes a draw. The parameters are referenced as \code{draw$parameter}.}
}
\description{
The parameters have their defined dimensions.
}
\examples{
map_draws(jags_fit, function(d) d$beta[1] + d$beta[2])
}
