% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{permute_draws}
\alias{permute_draws}
\title{Permute Draws}
\usage{
permute_draws(draws, within_chain = FALSE)
}
\arguments{
\item{draws}{Draw Data Frame as generated by \code{\link[=get_draws]{get_draws()}}.}

\item{within_chain}{Logical, whether to permute all draws (default) or permute
iterations within each chain.}
}
\description{
Draws taken from the MCMC algorithm are correlated with the previous and next draws.
Permute them to remove this effect.
}
