% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{run_cmdstan_advi}
\alias{run_cmdstan_advi}
\title{Fit Automatic Differentiation Variational Inference (ADVI) with CmdStan.}
\usage{
run_cmdstan_advi(
  model,
  file,
  data,
  inits = NULL,
  n_draws = 1000,
  algorithm = c("meanfield", "fullrank"),
  ...
)
}
\arguments{
\item{model}{CmdStanModel as generated by \code{\link[=cmdstan_model]{cmdstan_model()}} or \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{file}{Alternatively, a .stan file containing the model code.}

\item{data}{List containing the data for fitting. Must contain all the variables
on the \code{data} block of the stan model.}

\item{algorithm}{Algorithm used for fitting. Can be \code{"meanfield"} or \code{"fullrank"}.}

\item{...}{}
}
\description{
See \code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}.
}
\seealso{
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}
}
