% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{run_jags}
\alias{run_jags}
\title{Fit a JAGS model}
\usage{
run_jags(
  model,
  file,
  data,
  inits = NULL,
  drop_parameters = character(),
  iter = 2000,
  burnin = floor(iter/2),
  thin = 1,
  chains = 4,
  warmup = burnin,
  refresh = iter/100,
  seed,
  ...
)
}
\arguments{
\item{model}{Jags model as a string or created with \code{\link[=jags_model]{jags_model()}}.}

\item{file}{Alternatively, a file containing the JAGS code.}

\item{data}{List containing the data used to train the model.}

\item{inits}{Initial values for parameters. Accepts either:
\itemize{
\item Unnamed list of a named list, with the initial values for the parameters. All chains will start
with these values. For example: \cr
\code{inits = list(list(a = 2, b = 1:5))}.
\item List of named lists, where each element is a named list of values corresponding to a chain.
For example, with 2 chains: \cr
\code{inits = list(list(a = 2, b = 1:5), list(a = 3, b = 2:6))}
\item Data.frame, where each row corresponds to a chain and each column corresponds to a parameter.
You can use a \code{\link[tibble:tibble]{tibble::tibble()}} for multivariate parameters.
For example: with 2 chains: \cr
\code{inits = tibble::tibble(a = 2:3, b = list(1:5, 2:6))}
}}

\item{drop_parameters}{Character vector of parameters to exclude from final output. Saves
runtime memory.}

\item{iter}{Number of iterations, including burnin, per chain.}

\item{burnin}{Number of samples to burn per chain.}

\item{thin}{Positive integer specifying the period for saving samples.}

\item{chains}{Positive integer specifying the number of Markov chains.
The default is 4.}

\item{warmup}{Alias for \code{burnin}.}

\item{refresh}{Positive Integer, number of iterations between updates of the progress bars.}

\item{seed}{Seed to use for the RNG. In JAGS, specifying a seed
simply runs \code{set.seed(seed)}.}

\item{...}{Arguments passed on to \code{\link[R2jags:jags]{R2jags::jags()}}.}
}
\value{
An rjags fit.
}
\description{
Generate posterior samples from a JAGS model using Markov Chain Monte-Carlo.
}
\seealso{
\code{\link[R2jags:jags]{R2jags::jags()}}
}
