% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{[[.thomas_draw_df}
\alias{[[.thomas_draw_df}
\title{Extract a draw and reshape it to the dimensions of each parameter}
\usage{
\method{[[}{thomas_draw_df}(x, ...)
}
\arguments{
\item{x}{Data frame containing the draws.}

\item{...}{Indexes. You can select one row and any number columns.}
}
\value{
A named list where each parameter has the right dimension.
}
\description{
Extract a draw and reshape it to the dimensions of each parameter
}
\examples{
draws <- get_draws(fit)
draws[[1, ]]
draws[1, ]
}
