% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{traceplot}
\alias{traceplot}
\title{Plot chains to check convergence.}
\usage{
traceplot(fit, ..., .max_values = 4L)
}
\arguments{
\item{fit}{Bayesian fit.}

\item{...}{Parameters to select, passed to [dplyr::select()].}

\item{.max_values}{For multivariate parameters, number of different values to plot.}
}
\description{
Plot chains to check convergence.
}
