% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_draws}
\alias{get_draws}
\alias{get_samples}
\title{Get posterior samples}
\usage{
get_draws(fit, as = c("tibble", "list", "array"))

get_samples(fit, as = c("tibble", "list", "array"))
}
\arguments{
\item{fit}{STAN or JAGS fit.}
}
\value{
data.frame with columns for iteration, chain, and each parameter.
}
\description{
Get the posterior samples from a bayesian fit.
}
