% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_statistic}
\alias{get_statistic}
\title{Get a statistic for each parameter.}
\usage{
get_statistic(fit, fun = mean)
}
\arguments{
\item{fit}{A Bayesian Fit or a Draw Data Frame as generated by \code{\link[=get_draws]{get_draws()}}.}

\item{fun}{Function to apply to the draws of each parameter.}
}
\value{
Named list, where the names are the parameters.
}
\description{
Get a statistic for each parameter.
}
\seealso{
\code{\link[=get_means]{get_means()}}, \code{\link[=get_sd]{get_sd()}}
}
