% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{jags_model}
\alias{jags_model}
\title{Write a JAGS model with R's syntax highlighting.}
\usage{
jags_model(model, data, file)
}
\arguments{
\item{model}{Expression or string containing the code for the model block.}

\item{data}{Expression or string containing the code for the data block.}

\item{file}{String indicating where the model is written.}
}
\description{
Instead of writing the model on a string, write it inside
this function and get syntax highlighting and warnings.
Alternatively, specify the file where the model is written.
}
\details{
In order to truncate a distribution, you can use the pipe.
\itemize{
\item Use the magrittr pipe (recommended):  \code{a ~ dnorm(0, 1) \%>\% T(2, )}
\item Alternatively, the native pipe:  \code{a ~ dnorm(0, 1) |> T(2, )}.
}
}
\examples{
my_model <- jags_model({
    a ~ dexp(0.01)
    b ~ dexp(0.01)

    for (i in 1:n) {
        lambda[i] ~ dgamma(a, b)
        y[i] ~ dpois(lambda[i])
    }
})
}
