% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{run_cmdstan}
\alias{run_cmdstan}
\title{Fit a CmdStan model}
\usage{
run_cmdstan(
  model,
  file,
  data,
  inits = NULL,
  iter = 2000,
  burnin = floor(iter/2),
  thin = 1,
  chains = 4,
  warmup = burnin,
  seed = NULL,
  ...
)
}
\arguments{
\item{model}{CmdStanModel as generated by \code{\link[=cmdstan_model]{cmdstan_model()}} or \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{file}{Alternatively, a .stan file containing the model code.}

\item{data}{List containing the data for fitting. Must contain all the variables
on the \code{data} block of the stan model.}

\item{inits}{Initial values for parameters. Accepts either:
\itemize{
\item Named list, with the initial values for the parameters. All chains will start
with these values. For example: \cr
\code{inits = list(a = 2, b = 1:5)}.
\item List of lists, where each element is a named list of values corresponding to a chain.
For example, with 2 chains: \cr
\code{inits = list(list(a = 2, b = 1:5), list(a = 3, b = 2:6))}
\item Data.frame, where each row corresponds to a chain and each column corresponds to a parameter.
For example: with 2 chains: \cr
\code{inits = data.frame(a = 2:3, b = list(1:5, 2:6))}
\item Function. Takes an input \code{chain_id} and outputs a named list. For example: \cr
\code{inits = function(chain_id) { list(a = chain_id, b = rpois(5, 1.2)) }}
}}

\item{iter}{Number of iterations, including burnin, per chain.}

\item{burnin}{Number of samples to burn per chain.}

\item{thin}{Positive integer specifying the period for saving samples.}

\item{chains}{Positive integer specifying the number of Markov chains.
The default is 4.}

\item{warmup}{Alias for \code{burnin}.}

\item{seed}{Seed to use for the RNG.}

\item{...}{Arguments passed on to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}
}
\description{
Generate posterior samples from a CmdStanModel using Markov Chain Monte-Carlo.
}
