% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{run_jags}
\alias{run_jags}
\title{Run MCMC on a JAGS model}
\usage{
run_jags(
  model,
  file,
  data,
  inits = NULL,
  parameters,
  iter = 2000,
  burnin = floor(iter/2),
  thin = 1,
  chains = 4,
  warmup = burnin,
  seed,
  ...
)
}
\arguments{
\item{model}{Jags model as a string or created with [jags_model()].}

\item{file}{Alternatively, a file where the JAGS model is written.}

\item{data}{List containing the data used to train the model.}

\item{inits}{List or data.frame of initial values for parameters.
Must be either length 1 or the number of chains.
If a list, each value must be a named list containing the initial value
for each parameter.
If a data.frame, each row must contain the initial value for each parameter.}

\item{parameters}{Character vector of parameters to monitor.}

\item{iter}{Number of iterations, including burnin, per chain.}

\item{burnin}{Number of samples to burn per chain.}

\item{thin}{Positive integer specifying the period for saving samples.}

\item{chains}{Positive integer specifying the number of Markov chains.
The default is 4.}

\item{warmup}{Alias for burnin.}

\item{seed}{Seed to use for the RNG. In JAGS, specifying a seed
simply runs \code{set.seed(seed)}.}

\item{...}{Arguments passed on to [R2jags::jags()].}
}
\value{
An rjags fit.
}
\description{
Generate posterior samples from a JAGS model using Markov Chain Monte-Carlo.
}
