% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{traceplot}
\alias{traceplot}
\title{Plot chains to check convergence.}
\usage{
traceplot(fit, ..., .max_values = 4L)
}
\arguments{
\item{fit}{A Bayesian Fit or a Draw Data Frame as generated by \code{\link[=get_draws]{get_draws()}}.}

\item{...}{Parameters to select, passed to \code{\link[dplyr:select]{dplyr::select()}}.}

\item{.max_values}{For multivariate parameters, number of different values to plot.}
}
\description{
Plot chains to check convergence.
}
